requirements.txt
pyodbc>=5.0
python-dotenv>=1.0
openpyxl>=3.1
pytest>=8.0
anthropic>=0.40
